Fix gvr-events according to GVRViewSO pattern
* Don't call getApplication().registerView( ) at application.
* Move code from Activity to GVRf Main class
* Fix layout
* Use IViewEvents listener.

GearVRf-DCO-1.0-Signed-off-by: Ragner Magalhaes <ragner.n@samsung.com>
