+++ gvr-events/app/src/main/res/layout/activity_main.xml
@@ -1,22 +1,25 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical" >
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="wrap_content"
+    android:layout_height="wrap_content"
+    android:background="@android:color/background_light"
+    android:orientation="vertical"
+    android:padding="10dp">
 
     <LinearLayout
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:gravity="center"
         android:layout_weight="0"
-        android:orientation="horizontal" >
+        android:gravity="center"
+        android:orientation="horizontal">
 
         <Button
             android:id="@+id/button1"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:minWidth="0sp"
             android:minHeight="0sp"
+            android:minWidth="0sp"
             android:text="@string/button1"
             android:textSize="9sp" />
 
@@ -24,8 +27,8 @@
             android:id="@+id/button2"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:minWidth="0sp"
             android:minHeight="0sp"
+            android:minWidth="0sp"
             android:text="@string/button2"
             android:textSize="9sp" />
 
@@ -38,12 +41,12 @@
     </LinearLayout>
 
     <LinearLayout
-        android:paddingTop="5sp"
         android:id="@+id/listLayout"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_weight="0"
-        android:orientation="vertical">
+        android:orientation="vertical"
+        android:paddingTop="5sp">
 
         <TextView
             android:layout_width="wrap_content"
@@ -54,15 +57,14 @@
         <ListView
             android:id="@+id/listView"
             android:layout_width="match_parent"
-            android:layout_height="35dp">
-        </ListView>
+            android:layout_height="35dp"></ListView>
     </LinearLayout>
 
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="0dp"
-        android:layout_weight="1"
         android:layout_gravity="bottom"
+        android:layout_weight="1"
         android:gravity="bottom"
         android:orientation="vertical">
 
@@ -86,8 +88,8 @@
             android:id="@+id/listTextView"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:gravity="center"
             android:layout_gravity="bottom"
+            android:gravity="center"
             android:text="@string/listClicked"
             android:textSize="10sp" />
     </LinearLayout>
